fix(sync): resolve sync engine bugs from UAT
- Fix specta BigIntForbidden: change u64 → u32 for GitHub IDs, poll
  intervals, and device flow fields (JS can't safely represent u64)
- Add HTTP scope to Tauri capabilities for github.com and api.github.com
- Handle empty repos: initialize dispatch.json on first read_workspace
  when manifest is missing (freshly cloned empty GitHub repos)
- Guard git2 credential callback against infinite retry loops
- Handle missing FETCH_HEAD gracefully in pull (empty repos)
- Fix SyncStatusChip not re-rendering: subscribe to syncStatuses map
  instead of getStatus function reference
- Add fallback status reset in triggerPull/triggerSync catch blocks
  to prevent chip getting stuck on "syncing" forever

+++ src-tauri/src/commands/auth.rs
@@ -21,7 +21,7 @@ pub async fn initiate_login(
 pub async fn poll_login(
     app: tauri::AppHandle,
     device_code: String,
-    interval: u64,
+    interval: u32,
 ) -> Result<GitHubUser, String> {
     let access_token =
         device_flow::poll_for_token(device_flow::get_client_id(), &device_code, interval).await?;
